HCAMP-23rd/WEB/Challenges/Top-Secret/requirements.txt
Flask
flask-sqlalchemy
flask-migrate
flask_wtf
cryptography
sqlalchemy-utils
python-decouple
pymysql